Removed old Ghash implementation
- added test
