+++ tls-2pc-core/src/msgs/mod.rs
@@ -1,4 +1,2 @@
-#[cfg(feature = "ghash")]
-pub mod ghash;
 #[cfg(feature = "prf")]
 pub mod prf;
